Add files via upload
This repository contains a simple HTML file for a login page. The page has a form with fields for username and password, and a submit button to simulate the login process. Please note that this is a front-end-only form and does not perform any real authentication.

The form is styled with basic CSS to position the elements, and it can be opened directly in a web browser.

For more of my projects, visit [AndriaSys](https://andriasys.com/).